refactor(admin): change folder structure
