+++ apps/admin/src/routes/lab/components/lab-table.tsx
